[spi] enable multi-byte backfeed echo
This was a challenge as the mailboxing code in the default implementation was not synthesizable on the Ice40,
due to use of D Flip Flops with async set/reset.
There are likely to be a few race conditions here, but I can successfully read from Linux SPI dev a 100khz.