Implement SPI Transmit
Due to the limitations and possible incorrectness of the selected implementation, the core Byte sender had to be redesigned.
This implements a much simpler mechanism for crossing the clock domain. It also aleviated CIPO preload issues by mapping
the tx_byte register directly to the wire and tristates the pin. Given we are likely to have all TX data a priori this
should be fine. E.g. all TX Bytes are register reads. However this may cause issues as the next byte is started at the
falling edge without any clock delay. This trips up my logic analyser.